commit
commit

+++ client/player.js
@@ -18,11 +18,14 @@ class Player{
     this.pos = {x:0, y:0};
     this.matrix = null;
     this.score= 0;
+    this.gameActive = true; // Track if game is active
 
     this.reset();
     }
 
     move(dir){
+    if (!this.gameActive) return; // Don't move if game over
+    
     this.pos.x += dir;
     if(this.arena.collide(this)){
         this.pos.x -= dir;
@@ -40,9 +43,9 @@ class Player{
     (this.matrix[0].length / 2 | 0);
 
     if (this.arena.collide(this) && this.score > 0){
-        this.arena.clear();
-        this.score = 0;
-        this.tetris.updateScore(this.score);
+        // Game over - player topped out
+        console.log('Game over - topped out!');
+        this.gameActive = false; // Stop game
         
         // Emit game over event
         this.events.emit('game-over');
@@ -74,8 +77,9 @@ class Player{
     }
 }
 
-
     rotate(dir){
+    if (!this.gameActive) return;
+    
     const pos = this.pos.x;
     let offset = 1;
     this.rotateMatrix(this.matrix, dir);
@@ -93,24 +97,29 @@ class Player{
 }
 
     drop(){
+    if (!this.gameActive) return;
+    
     this.pos.y++;
     if(this.arena.collide(this)){
         this.pos.y--;
         this.arena.merge(this);
         this.reset();
-        
-        const sweepResult = this.arena.sweep();
-        this.score += sweepResult.score;
-        
-        // Emit score with lines cleared info
-        this.events.emit('score', this.score, sweepResult.linesCleared);
-        this.tetris.updateScore(this.score);
+
+        if (this.gameActive) {
+            const sweepResult = this.arena.sweep();
+            this.score += sweepResult.score;
+            
+            // Emit score with lines cleared info
+            this.events.emit('score', this.score, sweepResult.linesCleared);
+            this.tetris.updateScore(this.score);
+        }
     }
     this.dropCounter=0;
     this.events.emit('pos', this.pos);
 }
 
 update(deltaTime){
+    if (!this.gameActive) return;
     this.dropCounter += deltaTime;
     if(this.dropCounter > this.dropInterval){
         this.drop();
